pyproject.toml
[project]
name = "cfp-radar"
version = "0.1.0"
description = "AI-powered tech events and CFP tracker for CI/CD, DevOps, and Cloud Native communities"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai",
    "httpx",
    "beautifulsoup4",
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "python-dateutil",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "respx",
    "ruff",
    "mypy",
    "types-python-dateutil",
    "types-PyYAML",
]

[project.scripts]
cfp-radar = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",   # line length handled by formatter
    "SIM105", # contextlib.suppress - keep explicit try/except for clarity
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_equality = true
warn_redundant_casts = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "google.*",
    "bs4",
    "dateutil.*",
]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
